requirements.txt
tensorflow>=2.20.0
keras>=2.20.0
flask>=2.3.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
werkzeug>=2.3.0
